Rename drowidget.py to dro_widget.py
